Revert in/out amps for Pin Complexes again
Reverting what was done on 5fb9241a104aafc38f4ae0ea50bdb81c02de588a

Probably a alsa bug, not a bug in the script. Definition of Input/Output
amplifiers for Pin Complexes is ambiguous.  /proc/asound/card*/codec*
shows as if the Pin Complex had multiple input amplifiers, but this
doesn't seem to be true.

The graph for hp-samba-*-fullvol.txt demonstrates that the definition
is incorrect: Line In is shown as muted, if Amp-Out is connected
to the 0x0b mixer. With this patch, the graph is rendered correctly:
(output to Line In is muted, and input from Line In is at 0 dB).

Signed-off-by: Eduardo Habkost <ehabkost@mandriva.com>